refactor: move executions to package on one's own
