+++ jdbc/src/main/java/nextstep/jdbc/JdbcTemplate.java
@@ -5,6 +5,8 @@ import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.util.List;
 import javax.sql.DataSource;
+import nextstep.jdbc.execution.ListExecution;
+import nextstep.jdbc.execution.ObjectExecution;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
